Create box context menu, update menu modals

+++ src/components/menus/DeletePrompt.tsx
@@ -1,7 +1,9 @@
 import { removeBoxAlbumThunk, removeBoxArtistThunk, removeBoxPlaylistThunk, removeBoxTrackThunk } from "core/features/currentBoxDetail/currentBoxDetailSlice";
 import { setModalState } from "core/features/modal/modalSlice";
+import { deleteUserBoxThunk } from "core/features/userBoxes/userBoxesSlice";
 import { useAppDispatch } from "core/hooks/useAppDispatch";
 import { AppThunk } from "core/store/store";
+import { useHistory } from "react-router-dom";
 import { Album, Artist, Playlist, Track } from "../../core/types/interfaces";
 import styles from "./DeletePrompt.module.css";
 
@@ -9,42 +11,65 @@ type MusicData = Artist | Album | Track | Playlist;
 
 interface IProps {
   boxId: string
-  itemData: MusicData
+  itemData?: MusicData
+  deleteType: "Box" | "Item"
 }
 
-function DeletePrompt({itemData, boxId}: IProps) {
+function DeletePrompt({itemData, boxId, deleteType}: IProps) {
   const dispatch = useAppDispatch();
-  let deleteAction: (box: string, item: string) => AppThunk;
+  const history = useHistory();
+  let promptMessage: string;
+  let deleteItem: (box: string, item: string) => AppThunk;
 
-  switch(itemData.type){
-    case "artist":
-      deleteAction = removeBoxArtistThunk
-    break;
-    case "album":
-      deleteAction = removeBoxAlbumThunk
-    break;
-    case "track":
-      deleteAction = removeBoxTrackThunk
-    break;
-    case "playlist":
-      deleteAction = removeBoxPlaylistThunk
-    break;
-    default:
-      deleteAction = removeBoxArtistThunk
+  if (deleteType === "Item" && itemData){
+    switch(itemData.type){
+      case "artist":
+        deleteItem = removeBoxArtistThunk
+      break;
+      case "album":
+        deleteItem = removeBoxAlbumThunk
+      break;
+      case "track":
+        deleteItem = removeBoxTrackThunk
+      break;
+      case "playlist":
+        deleteItem = removeBoxPlaylistThunk
+      break;
+      default:
+        deleteItem = removeBoxArtistThunk
+    }
   }
 
-  const handleDeleteItem = () => {
-    dispatch(deleteAction(boxId, itemData._id!))
+  switch(deleteType){
+    case "Box":
+      promptMessage = 'Are you sure you want to delete this box from your collection?'
+    break;
+    case "Item":
+      promptMessage = 'Are you sure you want to remove this item from the box?'
+    break;
+    default:
+      promptMessage = ''
+  }
+
+  const handleDeleteAction = () => {
+    if (deleteType === "Box"){
+      console.log('Delete action')
+      dispatch(deleteUserBoxThunk(boxId));
+      history.push('/')
+    }
+    else if (deleteType === "Item" && itemData){
+      dispatch(deleteItem(boxId, itemData._id!));
+    }
     dispatch(setModalState({visible: false, type:"", boxId:"", page: "", itemData: undefined}))
   }
 
   return (
     <div id={styles.modalBody}>
       <div id={styles.confirmation}>
-        <h3> Are you sure you want to remove this item from the box? </h3>
+        <h3> {promptMessage} </h3>
       </div>
       <div id={styles.modalFooter}>
-        <button onClick={() => handleDeleteItem()}> Yes, delete item </button>
+        <button onClick={() => handleDeleteAction()}> Yes, delete it </button>
         <button onClick={() => dispatch(setModalState({visible: false, type:"", boxId:"", page: "", itemData: undefined}))}> Cancel </button>
       </div>
     </div>
